# docs/public-api/references/swagger/form/flowable-swagger-form.yaml
...
swagger: "2.0"
info:
  description: "# flowable / flowəb(ə)l /\r\n\r\n- a compact and highly efficient\
    \ workflow and Business Process Management (BPM) platform for developers, system\
    \ admins and business users.\r\n- a lightning fast, tried and tested BPMN 2 process\
    \ engine written in Java. It is Apache 2.0 licensed open source, with a committed\
    \ community.\r\n- can run embedded in a Java application, or as a service on a\
    \ server, a cluster, and in the cloud. It integrates perfectly with Spring. With\
    \ a rich Java and REST API, it is the ideal engine for orchestrating human or\
    \ system activities."
  version: "v1"
  title: "Flowable FORM REST API"
  contact:
    name: "Flowable"
    url: "http://www.flowable.org/"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
basePath: "/flowable-rest/form-api"
...
schemes:
- "http"
- "https"
paths:
...
  /form-repository/deployments:
    get:
      tags:
      - "Form Deployments"
      summary: "List of Form Deployments"
      description: ""
      operationId: "listFormDeployments"
      produces:
      - "application/json"
      parameters:
      - name: "name"
        in: "query"
        description: "Only return form deployments with the given name."
        required: false
        type: "string"
      - name: "nameLike"
        in: "query"
        description: "Only return form deployments with a name like the given name."
        required: false
        type: "string"
      - name: "category"
        in: "query"
        description: "Only return form deployments with the given category."
        required: false
        type: "string"
      - name: "categoryNotEquals"
        in: "query"
        description: "Only return form deployments which do not have the given category."
        required: false
        type: "string"
      - name: "parentDeploymentId"
        in: "query"
        description: "Only return form deployments with the given parent deployment\
          \ id."
        required: false
        type: "string"
      - name: "parentDeploymentIdLike"
        in: "query"
        description: "Only return form deployments with a parent deployment id like\
          \ the given value."
        required: false
        type: "string"
      - name: "tenantId"
        in: "query"
        description: "Only return form deployments with the given tenantId."
        required: false
        type: "string"
      - name: "tenantIdLike"
        in: "query"
        description: "Only return form deployments with a tenantId like the given\
          \ value."
        required: false
        type: "string"
      - name: "withoutTenantId"
        in: "query"
        description: "If true, only returns form deployments without a tenantId set.\
          \ If false, the withoutTenantId parameter is ignored."
        required: false
        type: "boolean"
      - name: "sort"
        in: "query"
        description: "Property to sort on, to be used together with the order."
        required: false
        type: "string"
        enum:
        - "id"
        - "name"
        - "deployTime"
        - "tenantId"
      responses:
        200:
          description: "Indicates the request was successful."
          schema:
            $ref: "#/definitions/DataResponseFormDeploymentResponse"
      security:
      - basicAuth: []
    post:
      tags:
      - "Form Deployments"
      summary: "Create a new form deployment"
      description: "The request body should contain data of type multipart/form-data.\
        \ Every file part is deployed: a file ending in .form or .xml is one form\
        \ definition, and a file ending in .zip or .bar is unpacked and each .form\
        \ or .xml entry inside it is deployed. The deployment name is the deploymentName\
        \ form field when given, otherwise the name of the (first) file."
      operationId: "uploadDeployment"
      consumes:
      - "multipart/form-data"
      produces:
      - "application/json"
      parameters:
      - name: "tenantId"
        in: "query"
        required: false
        type: "string"
      - name: "file"
        in: "formData"
        description: "A .form or .xml definition, or a .zip/.bar holding several.\
          \ Repeat for several files."
        required: false
        type: "file"
      - name: "deploymentName"
        in: "formData"
        description: "Name of the deployment. Defaults to the file name."
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/FormDeploymentResponse"
        201:
          description: "Indicates the form deployment was created."
        400:
          description: "Indicates there was no content present in the request body\
            \ or the content mime-type is not supported for form deployment. The status-description\
            \ contains additional information."
      security:
      - basicAuth: []
...
securityDefinitions:
  basicAuth:
    type: "basic"
definitions:
...
  DataResponseFormDeploymentResponse:
    type: "object"
    properties:
      data:
        type: "array"
        items:
          $ref: "#/definitions/FormDeploymentResponse"
      total:
        type: "integer"
        format: "int64"
      start:
        type: "integer"
        format: "int32"
      sort:
        type: "string"
      order:
        type: "string"
      size:
        type: "integer"
        format: "int32"
...
  FormDeploymentResponse:
    type: "object"
    properties:
      id:
        type: "string"
        example: "10"
      name:
        type: "string"
        example: "flowable-form-examples"
      deploymentTime:
        type: "string"
        format: "date-time"
        example: "2010-10-13T14:54:26.750+02:00"
      category:
        type: "string"
        example: "examples"
      url:
        type: "string"
        example: "http://localhost:8081/form-api/form-repository/deployments/10"
      parentDeploymentId:
        type: "string"
        example: "5"
      tenantId:
        type: "string"
        example: "null"
...